perf: f64→f32 embeddings, parking_lot mutex, busy handler, SSE fix
- Embeddings stored as f32 (4 bytes) instead of f64 (8 bytes): halves
  memory footprint for vector index. Auto-detects legacy f64 blobs
  (12288 bytes = 1536×8) and down-converts on read.
- Replaced std::sync::Mutex with parking_lot::Mutex for EmbedCache:
  no poison risk, slightly faster.
- Added r2d2 ConnectionCustomizer with 5s busy_timeout on every pooled
  connection: prevents SQLITE_BUSY under concurrent write pressure.
- SSE parsing no longer relies on first line being 'data:' — scans all
  lines for data: prefix, handles leading comment/event/empty lines.
- FTS rebuild detection samples 5 CJK memories instead of 1 to reduce
  false-positive rebuilds on startup.

+++ src/api/recall_handlers.rs
@@ -327,20 +327,20 @@ pub(super) async fn do_resume(
             } else {
                 // Compute centroid of context embeddings
                 let dim = context_embeds[0].1.len();
-                let mut centroid = vec![0.0f64; dim];
+                let mut centroid = vec![0.0f32; dim];
                 for (_, emb) in &context_embeds {
                     for (i, v) in emb.iter().enumerate() {
                         if i < dim { centroid[i] += v; }
                     }
                 }
-                let n = context_embeds.len() as f64;
+                let n = context_embeds.len() as f32;
                 for v in &mut centroid {
                     *v /= n;
                 }
 
                 let core_ids: Vec<String> = core.iter().map(|m| m.id.clone()).collect();
                 let core_embeds = d.get_embeddings_by_ids(&core_ids);
-                let core_embed_map: std::collections::HashMap<&str, &Vec<f64>> = core_embeds.iter()
+                let core_embed_map: std::collections::HashMap<&str, &Vec<f32>> = core_embeds.iter()
                     .map(|(id, emb)| (id.as_str(), emb))
                     .collect();
